# docs/api.yaml
openapi: 3.0.3
info:
  title: FollowingFeed API
  version: 1.0.0
  description: >-
    FollowingFeed 后端接口契约。除公开接口外，请通过 Bearer Token 认证。
    成功返回 2xx，客户端错误返回 4xx，服务端错误返回 5xx；
    响应体中的 code 用于细分业务结果。除健康检查和文档路由外，业务接口均受
    Redis IP 限流保护：超限返回空响应体的 HTTP 429，限流依赖异常返回空响应体的 HTTP 500。
    需要 Bearer 认证的接口在 Redis 会话检查失败时返回空响应体的 HTTP 503。
servers:
  - url: /api/v1
...
paths:
...
  /articles/edit:
    post:
      tags: [Article]
      summary: 保存文章草稿
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleRequest" }
      responses:
        "200": { $ref: "#/components/responses/ArticleIDResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLargeResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /articles/publish:
    post:
      tags: [Article]
      summary: 发布文章
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleRequest" }
      responses:
        "200": { $ref: "#/components/responses/ArticleIDResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLargeResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /articles/withdraw:
    post:
      tags: [Article]
      summary: 撤回已发布文章
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleIdRequest" }
      responses:
        "200": { $ref: "#/components/responses/ArticleIDResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLargeResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /articles/delete:
    post:
      tags: [Article]
      summary: 软删除文章
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArticleIdRequest" }
      responses:
        "200": { $ref: "#/components/responses/ArticleIDResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "413": { $ref: "#/components/responses/PayloadTooLargeResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /articles/list:
    get:
      tags: [Article]
      summary: 获取当前用户的文章列表
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/OwnArticlePageResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /articles/detail/{id}:
    get:
      tags: [Article]
      summary: 获取当前用户的一篇文章
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200": { $ref: "#/components/responses/ArticleResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "403": { $ref: "#/components/responses/ForbiddenResponse" }
        "404": { $ref: "#/components/responses/NotFoundResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/list:
    get:
      tags: [Article]
      summary: 获取公开文章列表
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/ArticlePageResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/detail/{id}:
    get:
      tags: [Article]
      summary: 获取公开文章详情
      parameters:
        - $ref: "#/components/parameters/ArticleId"
      responses:
        "200": { $ref: "#/components/responses/ArticleResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "404": { $ref: "#/components/responses/NotFoundResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /feed:
    get:
      tags: [Article]
      summary: 获取关注作者的文章 Feed
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/ArticlePageResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "401": { $ref: "#/components/responses/UnauthorizedResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
  /pub/users/{id}/articles:
    get:
      tags: [Article]
      summary: 获取指定作者的已发布文章
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200": { $ref: "#/components/responses/ArticlePageResultResponse" }
        "400": { $ref: "#/components/responses/BadRequestResponse" }
        "500": { $ref: "#/components/responses/InternalServerErrorResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: followingfeed_refresh_token
  parameters:
    UserId:
      name: id
      in: path
      required: true
      description: 用户 ID
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
    ArticleId:
      name: id
      in: path
      required: true
      description: 文章 ID
      schema: { type: integer, format: int64, minimum: 1 }
      example: 1
    Page:
      name: page
      in: query
      required: false
      description: 页码，从 1 开始，默认 1
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      required: false
      description: 每页数量，默认 10，最大 50
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
  schemas:
...
    ArticleRequest:
      type: object
      required: [title, content]
      properties:
        id:
          type: integer
          format: int64
          minimum: 0
          default: 0
          description: 新建传 0；编辑或发布已有文章时传文章 ID
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: 最多 200 个 Unicode 字符
          example: Go 服务依赖注入实践
        content:
          type: string
          minLength: 1
          description: >-
            非空 UTF-8 正文，编码后最大 60 KiB（61440 字节）。OpenAPI 3.0
            的 maxLength 按字符计数，无法准确表达字节上限，因此不声明 maxLength。
          example: 文章正文
    ArticleIdRequest:
      type: object
      required: [id]
      properties:
        id: { type: integer, format: int64, minimum: 1, example: 1 }
...
    ResultMeta:
      type: object
      required: [code, msg]
      properties:
        code: { type: integer, example: 0 }
        msg: { type: string, example: success }
...
    BusinessErrorResult:
      type: object
      required: [code, msg, data]
      properties:
        code:
          type: integer
          enum: [4, 5]
          description: 4 表示参数或权限等业务错误，5 表示服务处理失败
          example: 4
        msg: { type: string, example: 参数错误 }
        data: { type: object, nullable: true, example: null }
...
    ArticleVO:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        title: { type: string, example: Go 服务依赖注入实践 }
        abstract: { type: string, description: 列表返回的正文摘要 }
        content: { type: string, description: 详情返回的完整正文 }
        authorId: { type: integer, format: int64, example: 1 }
        authorNickname: { type: string, example: 云端旅人 }
        status:
          {
            type: integer,
            format: int32,
            enum: [0, 1, 2, 3],
            description: 0未知、1草稿、2已发布、3私密,
          }
        createdAt:
          {
            type: integer,
            format: int64,
            description: Unix 毫秒时间戳,
            example: 1786075200000,
          }
        updatedAt:
          {
            type: integer,
            format: int64,
            description: Unix 毫秒时间戳,
            example: 1786075200000,
          }
    ArticlePageVO:
      type: object
      required: [items, page, pageSize, total, totalPages]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ArticleVO" }
        page: { type: integer, minimum: 1, example: 1 }
        pageSize: { type: integer, minimum: 1, maximum: 50, example: 10 }
        total: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    OwnArticlePageVO:
      type: object
      required: [list, summary]
      properties:
        list: { $ref: "#/components/schemas/ArticlePageVO" }
        summary:
          type: object
          required: [draft, published]
          properties:
            draft: { type: integer, format: int64, minimum: 0 }
            published: { type: integer, format: int64, minimum: 0 }
    ArticleResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/ArticleVO" }
    ArticleIDResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data:
              type: integer
              format: int64
              minimum: 1
              description: 新建、更新或状态变更后的文章 ID
    ArticlePageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/ArticlePageVO" }
    OwnArticlePageResult:
      allOf:
        - $ref: "#/components/schemas/ResultMeta"
        - type: object
          required: [data]
          properties:
            data: { $ref: "#/components/schemas/OwnArticlePageVO" }
...
  responses:
...
    ArticleResultResponse:
      description: 文章详情或业务错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ArticleResult" } },
        }
    ArticleIDResultResponse:
      description: 文章写入或状态变更成功，data 为文章 ID
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ArticleIDResult" }
    ArticlePageResultResponse:
      description: 文章分页结果或业务错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/ArticlePageResult" } },
        }
    OwnArticlePageResultResponse:
      description: 当前用户文章分页及状态统计
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/OwnArticlePageResult" } },
        }
...
    BadRequestResponse:
      description: 请求格式或参数错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    PayloadTooLargeResponse:
      description: 请求体超过 1 MiB 限制
      content:
        application/json:
          schema: { $ref: "#/components/schemas/BusinessErrorResult" }
    ForbiddenResponse:
      description: 已认证，但没有资源访问权限
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    NotFoundResponse:
      description: 请求的资源不存在
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
...
    InternalServerErrorResponse:
      description: 服务内部错误
      content:
        {
          application/json:
            { schema: { $ref: "#/components/schemas/BusinessErrorResult" } },
        }
    UnauthorizedResponse:
      description: Token 缺失、无效或会话已失效；响应体为空